# chore: makes the test a bit faster